COMP: Eliminate compiler warnings on the 64-bit Windows dashboards.

+++ VTK/Common/vtkDirectory.cxx
@@ -265,11 +265,11 @@ int vtkDirectory::FileIsDirectory(const char *name)
   int n = 0;   
   if (!absolutePath && this->Path)   
     {   
-    n = strlen(this->Path);   
+    n = static_cast<int>(strlen(this->Path));
     }   
    
-  int m = strlen(name);   
-   
+  int m = static_cast<int>(strlen(name));
+
   fullPath = new char[n+m+2];   
    
   if (!absolutePath && this->Path)
