+++ VTK/Common/vtkWin32ProcessOutputWindow.cxx
@@ -161,7 +161,7 @@ void vtkWin32ProcessOutputWindow::Write(const char* data, size_t length)
 
     // Write the data to the pipe.  If it breaks, close the pipe.
     DWORD nWritten;
-    if(!WriteFile(this->OutputPipe, data, length, &nWritten, 0))
+    if(!WriteFile(this->OutputPipe, data, (DWORD) length, &nWritten, 0))
       {
       this->Broken = 1;
       CloseHandle(this->OutputPipe);
