+++ VTK/Graphics/vtkAssignAttribute.cxx
@@ -49,7 +49,7 @@ vtkAssignAttribute::vtkAssignAttribute()
     {
     for (int i = 0; i < vtkDataSetAttributes::NUM_ATTRIBUTES; i++)
       {
-      int l = strlen(vtkDataSetAttributes::GetAttributeTypeAsString(i));
+      int l = static_cast<int>(strlen(vtkDataSetAttributes::GetAttributeTypeAsString(i)));
       for (int c = 0; c < l && c < 19; c++)
         {
         vtkAssignAttribute::AttributeNames[i][c] =
