+++ VTK/Graphics/vtkMaskFields.cxx
@@ -44,7 +44,7 @@ vtkMaskFields::vtkMaskFields()
     {
     for (int i = 0; i < vtkDataSetAttributes::NUM_ATTRIBUTES; i++)
       {
-      int l = strlen(vtkDataSetAttributes::GetAttributeTypeAsString(i));
+      int l = static_cast<int>(strlen(vtkDataSetAttributes::GetAttributeTypeAsString(i)));
       for (int c = 0; c < l && c < 10; c++)
         {
         vtkMaskFields::AttributeNames[i][c] =
