+++ VTK/Hybrid/vtkDSPFilterGroup.cxx
@@ -298,7 +298,7 @@ void vtkDSPFilterGroup::Copy( vtkDSPFilterGroup *other )
 //----------------------------------------------------------------------------
 int vtkDSPFilterGroup::GetNumFilters( ) 
 {
-  return this->FilterDefinitions->m_vector.size();
+  return static_cast<int>(this->FilterDefinitions->m_vector.size());
 }
